Source_Code/requirements.txt
alembic==1.13.2
arrow==1.3.0
bcrypt==4.1.3
binaryornot==0.4.4
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
environs==11.0.0
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-DebugToolbar==0.15.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.4.1
Flask-WTF==1.2.1
fonttools==4.53.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
repoze.lru==0.7
requests==2.32.3
rich==13.7.1
Routes==2.5.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sklearn-preprocessing==0.1.0
SQLAlchemy==2.0.31
supervisor==4.2.5
text-unidecode==1.3
threadpoolctl==3.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
WTForms==3.1.2
zope.event==5.0
zope.interface==6.4.post2
